Fix build warnings.

+++ Signal/src/ViewControllers/ConversationView/ConversationViewController.m
@@ -3750,8 +3750,9 @@ typedef enum : NSUInteger {
             lastVisibleIndexPath = indexPath;
         }
     }
-    if (lastVisibleIndexPath && lastVisibleIndexPath.row >= self.viewItems.count) {
-        return (self.viewItems.count > 0 ? [NSIndexPath indexPathForRow:self.viewItems.count - 1 inSection:0] : nil);
+    if (lastVisibleIndexPath && lastVisibleIndexPath.row >= (NSInteger)self.viewItems.count) {
+        return (self.viewItems.count > 0 ? [NSIndexPath indexPathForRow:(NSInteger)self.viewItems.count - 1 inSection:0]
+                                         : nil);
     }
     return lastVisibleIndexPath;
 }
